refactor(qs): deduplicate tile services with BaseSettingTileService
