+++ app/src/main/java/com/shubhang/loophole/service/UsbDebugTileService.kt
@@ -1,78 +1,13 @@
 package com.shubhang.loophole.service
 
-import android.annotation.SuppressLint
-import android.app.PendingIntent
-import android.content.Intent
-import android.os.Build
-import android.service.quicksettings.Tile
-import android.service.quicksettings.TileService
 import com.shubhang.loophole.R
-import com.shubhang.loophole.appContainer
 import com.shubhang.loophole.settings.SecureSetting
-import com.shubhang.loophole.settings.SettingsWriteResult
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.launch
 
 /**
  * Quick Settings tile that toggles USB Debugging.
  */
-class UsbDebugTileService : TileService() {
-
-    private val devSettings by lazy { appContainer.devSettings }
-
-    private val serviceScope = CoroutineScope(SupervisorJob() + Dispatchers.Main.immediate)
-    private var listeningJob: Job? = null
-
-    override fun onStartListening() {
-        super.onStartListening()
-        listeningJob = serviceScope.launch {
-            devSettings.isUsbDebuggingEnabled.collect(::renderTile)
-        }
-    }
-
-    override fun onStopListening() {
-        listeningJob?.cancel()
-        listeningJob = null
-        super.onStopListening()
-    }
-
-    override fun onDestroy() {
-        serviceScope.cancel()
-        super.onDestroy()
-    }
-
-    override fun onClick() {
-        super.onClick()
-        serviceScope.launch {
-            if (devSettings.toggle(SecureSetting.USB_DEBUGGING) is SettingsWriteResult.PermissionDenied) {
-                openApp()
-            }
-        }
-    }
-
-    private fun renderTile(enabled: Boolean) {
-        val tile = qsTile ?: return
-        tile.state = if (enabled) Tile.STATE_ACTIVE else Tile.STATE_INACTIVE
-        tile.label = getString(if (enabled) R.string.tile_usb_label_on else R.string.tile_usb_label_off)
-        tile.updateTile()
-    }
-
-    @SuppressLint("StartActivityAndCollapseDeprecated")
-    private fun openApp() {
-        val launch = packageManager.getLaunchIntentForPackage(packageName)
-            ?.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK) ?: return
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
-            val pending = PendingIntent.getActivity(
-                this, 0, launch, PendingIntent.FLAG_IMMUTABLE
-            )
-            startActivityAndCollapse(pending)
-        } else {
-            @Suppress("DEPRECATION")
-            startActivityAndCollapse(launch)
-        }
-    }
-}
+class UsbDebugTileService : BaseSettingTileService(
+    setting = SecureSetting.USB_DEBUGGING,
+    labelOnRes = R.string.tile_usb_label_on,
+    labelOffRes = R.string.tile_usb_label_off,
+)
